Refactor NotificationComponent and NotificationComponentSecure to use Spring context for dependency injection. Update NotificationComponentBean constructor to accept NotificationClientService as a parameter. Introduce ThemeToggle component for theme switching in the UI. Enhance NotificationView to sort notifications by creation date and improve notification handling logic. Update application properties for JWT secret key configuration.

+++ src/main/java/org/dasher/speed/taskmanagement/repository/AppointmentRepository.java
@@ -76,5 +76,5 @@ public interface AppointmentRepository extends JpaRepository<Appointment, Intege
            "LEFT JOIN FETCH a.person_doctor pd " +
            "LEFT JOIN FETCH a.person_patient pp " +
            "WHERE a.id = :id")
-    Optional<Appointment> findByIdWithDetails(@Param("id") Integer id);
+    Optional<Appointment> findByIdWithDetails(@Param("id") long id);
 }
